review(todo-app): resplit task/model
