--- todo-app/src/entities/task/model/events.ts
@@ -1,5 +1,0 @@
-import { createEvent } from "effector";
-import type { QueryConfig } from "./types";
-
-export const toggleTask = createEvent<number>();
-export const setQueryConfig = createEvent<QueryConfig>();
