--- todo-app/src/entities/task/model/types.ts
@@ -1,4 +1,0 @@
-export type QueryConfig = {
-    completed?: boolean;
-    userId?: number;
-};
